Move model & dataset to separate directories
