+++ mask_detector/models/basic_cnn.py
